+++ src/main/java/org/geowe/service/filter/DivisionFilterImpl.java
@@ -18,11 +18,13 @@ package org.geowe.service.filter;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerRequestFilter;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.PathSegment;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.ext.Provider;
@@ -31,13 +33,11 @@ import org.apache.commons.io.IOUtils;
 import org.geowe.service.geometry.engine.JTSGeoEngineerHelper;
 import org.geowe.service.messages.Messages;
 import org.geowe.service.messages.Messages.Bundle;
-import org.geowe.service.model.FlatGeometry;
-import org.geowe.service.model.OperationData;
+import org.geowe.service.model.DivisionData;
 import org.geowe.service.model.error.ErrorEntityFactory;
 import org.jboss.logging.Logger;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.vividsolutions.jts.geom.LineString;
 
 /**
  * Filter for validate input data for division. OperationData.source: must be a
@@ -53,42 +53,57 @@ public class DivisionFilterImpl implements ContainerRequestFilter {
 	private final Logger log = Logger.getLogger(DivisionFilterImpl.class);
 	private Messages errorMessages;
 
+	JTSGeoEngineerHelper helper;
 	@Context
 	private HttpServletRequest request;
 
 	@Override
 	public void filter(ContainerRequestContext context) throws IOException {
 		String json = IOUtils.toString(context.getEntityStream(), "UTF-8");
+		
+		DivisionData opData = getBody(json);
 
-		OperationData opData = getBody(json);
-
-		FlatGeometry flatGeometry = opData.getOverlayData().iterator().next();
-		JTSGeoEngineerHelper helper = new JTSGeoEngineerHelper();
-		if (!(helper.getGeom(flatGeometry.getWkt()) instanceof LineString)) {
-			buildErrorResponse(context);
+		String wktdivisionLine = opData.getWktDivisionLine();
+		helper = new JTSGeoEngineerHelper();
+		if (!isWktLineString(wktdivisionLine)) {
+			buildErrorResponse(context, "overlay.must.be.line");
 		}
-
+		if(request.getPathInfo().contains("polygon") && !isWktPolygon(opData.getWktToDivide())){
+			buildErrorResponse(context, "overlay.must.be.polygon");
+		}
+		if(request.getPathInfo().contains("line") && !isWktLineString(opData.getWktToDivide())){
+			buildErrorResponse(context, "overlay.must.be.line");
+		}
+		
 		InputStream is = new ByteArrayInputStream(json.getBytes());
 		context.setEntityStream(is);
 
 	}
 
-	private OperationData getBody(String json) throws IOException {
+	private DivisionData getBody(String json) throws IOException {
 		ObjectMapper mapper = new ObjectMapper();
-		OperationData opData = mapper.readValue(json, OperationData.class);
+		DivisionData opData = mapper.readValue(json, DivisionData.class);
 
 		return opData;
 	}
 
-	private void buildErrorResponse(ContainerRequestContext context) {
+	private void buildErrorResponse(ContainerRequestContext context, String msgKey) {
 		ErrorEntityFactory errorEntityFactory = new ErrorEntityFactory(request.getLocale());
 		errorMessages = new Messages(Bundle.ERRORS, request.getLocale());
 		context.abortWith(
 				Response.status(Status.BAD_REQUEST)
 						.entity(errorEntityFactory.getBadRequestError(
 								new IllegalArgumentException(
-										errorMessages.getMessage("overlay.must.be.line"))))
+										errorMessages.getMessage(msgKey))))
 						.build());
+	}
+	
+	private boolean isWktPolygon(String wkt){
+		return helper.getGeom(wkt).getGeometryType().equals("Polygon");
+	}
+	
+	private boolean isWktLineString(String wkt){
+		return helper.getGeom(wkt).getGeometryType().equals("LineString");
 	}
 
 }