+++ src/main/java/org/geowe/service/geometry/engine/GeoEngineer.java
@@ -18,6 +18,7 @@ package org.geowe.service.geometry.engine;
 import java.util.Collection;
 import java.util.List;
 
+import org.geowe.service.model.DivisionData;
 import org.geowe.service.model.FlatGeometry;
 import org.geowe.service.model.OperationData;
 
@@ -184,17 +185,13 @@ public interface GeoEngineer {
 	 *            polygons to divide. operationData.overlay: division line
 	 * @return
 	 */
-	List<String> dividePolygons(OperationData operationData);
-
+	List<String> dividePolygon(DivisionData divisionData);
+	
 	/**
-	 * Divide a collection of LineStrings in source data with a line in overlay
-	 * data
-	 * 
-	 * @param operationData
-	 *            Operation data to calculate division. operationData.source:
-	 *            lineStrings to divide.  operationData.overlay: division line
+	 * Divide a LineString
+	 * @param divisionData
 	 * @return
 	 */
-	List<String> divideLines(OperationData operationData);
+	List<String> divideLine(DivisionData divisionData);
 
 }